% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_conv.R
\name{ar_conv}
\alias{ar_conv}
\title{Convergence object of an object of class `ar`.}
\usage{
ar_conv(object)
}
\arguments{
\item{object}{an object of class `ar`.}
}
\value{
\describe{\code{ar_conv} returns an object of class "ar_conv". An object of class "ar_conv" is a list containing the following components:
  \item{acceptance_rate_cumulative}{Ratio of accepted \code{x}, cumulative.}
  \item{y}{Accepted values \code{y}.}
  \item{mean_cumulative}{Sample mean of accepted values, cumulative.}
  \item{var_cumulative}{Sample variance of accepted values, cumulative.}
\describe{These components are used for the convergence plot using the function "ar_plot()" or "plot()."}
}
}
\description{
Provide a convergence object for an object of class `ar`.
}
