% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_plot.R
\name{ar_plot}
\alias{ar_plot}
\title{Plots of `ar` and `ar_conv`.}
\usage{
ar_plot(model, which = c(1, 2, 3, 4))
}
\arguments{
\item{model}{an object of class `ar` or `ar_conv`.}

\item{which}{index of convergence plot to be shown. Consists of index 1 (acceptance rate), 2 (y values), 3 (mean), and 4 (variance).}
}
\value{
An overlay plot of the target density f(x) and proposal density q(x) for an object of class `ar`, or convergence plots for an object of class `ar_conv`.
}
\description{
Display an overlay plot of the target density f(x) and proposal density q(x) for an object of class `ar`, or convergence plots for an object of class `ar_conv`.
}
\examples{
# pair of standard normal distribution (target) and uniform distribution (proposal)
f_norm <- "(1/sqrt(2*pi*1^2))*exp((-(x-0)^2)/(2*1^2))"
q_unif <- "runif(n, -4, 4)"
example_norm_unif <- ar(f_norm, q_unif, c=3.2, n=1000)
ar_plot(example_norm_unif)

}
