% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_print.R
\name{ar_print}
\alias{ar_print}
\title{Data of an object of class `ar`}
\usage{
ar_print(object)
}
\arguments{
\item{x}{an object of class `ar`.}

\item{...}{additional arguments.}
}
\value{
the data of an object of class `ar` with the following components: `x`, `f`, `q`, `y`, `r`, and `u`.
}
\description{
Show the data for an object of class `ar`.
}
