% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_summary.R
\name{ar_summary}
\alias{ar_summary}
\title{Summaries of an object of class `ar`}
\usage{
ar_summary(object)
}
\arguments{
\item{object}{an object of class `ar`.}
}
\value{
The summary of an object of class `ar` with the following components: `acceptance rate`, `empirical expected value of target distribution`, `empirical variance of target distribution`, `suggested minimum c`.
}
\description{
Provide a summary for an object of class `ar`.
}
