% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ar}
\alias{print.ar}
\title{Data of an object of class `ar`}
\usage{
\method{print}{ar}(x, ...)
}
\arguments{
\item{x}{an object of class `ar`.}

\item{...}{additional arguments.}
}
\value{
the data of an object of class `ar` with the following components: `x`, `f`, `q`, `y`, `r`, and `u`.
}
\description{
Show the data for an object of class `ar`.
}
