% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ar_conv}
\alias{print.ar_conv}
\title{Convergence data of an object of class `ar_conv`}
\usage{
\method{print}{ar_conv}(x, ...)
}
\arguments{
\item{x}{an object of class `ar_conv`.}

\item{...}{additional arguments.}
}
\value{
the data of an object of class `ar_conv` with the following components: `acceptance_rate_cumulative`, `y`, `mean_cumulative`, and `var_cumulative`.
}
\description{
Show the data for an object of class `ar_conv`.
}
