% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ar}
\alias{summary.ar}
\title{Summaries of an object of class `ar`}
\usage{
\method{summary}{ar}(object, ...)
}
\arguments{
\item{object}{an object of class `ar`.}

\item{...}{additional arguments.}
}
\value{
The summary of an object of class `ar` with the following components: `acceptance rate`, `empirical expected value of target distribution`, `empirical variance of target distribution`, `suggested minimum c`.
}
\description{
Provide a summary for an object of class `ar`.
}
