% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ar_conv}
\alias{summary.ar_conv}
\title{Summaries of an object of class `ar_conv`}
\usage{
\method{summary}{ar_conv}(object, ...)
}
\arguments{
\item{object}{an object of class `ar_conv`.}

\item{...}{additional arguments.}
}
\value{
The summary of an object of class `ar_conv` with the following components: `acceptance rate`, `empirical expected value of target distribution`, `empirical variance of target distribution`.
}
\description{
Provide a summary for an object of class `ar_conv`.
}
