% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_conv.R
\name{ar_conv}
\alias{ar_conv}
\title{Convergence diagnostics of an object of class `ar`.}
\usage{
ar_conv(object)
}
\arguments{
\item{object}{an object of class `ar`.}
}
\value{
The convergence diagnostic of an object of class `ar` with the following components: `acceptance rate`, `empirical expected value of target distribution`, `empirical variance of target distribution`, `suggested minimum c`.
}
\description{
Provide a convergence diagnostic for an object of class `ar`.
}
