% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_plot.R
\name{ar_plot}
\alias{ar_plot}
\title{ar_plot}
\usage{
ar_plot(model)
}
\arguments{
\item{model}{an object of class `ar`.}
}
\value{
An overlay plot of the target density f(x) and proposal density q(x).
}
\description{
Display an overlay plot of the target density f(x) and proposal density q(x).
}
