% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_print.R
\name{ar_print}
\alias{ar_print}
\title{Data of an object of class `ar` or `ar_conv`.}
\usage{
ar_print(object)
}
\arguments{
\item{object}{an object of class `ar` or `ar_conv`.}
}
\value{
\code{ar_print} returns different results depending on the class of the input object. If the input object is `ar`, this function will return a list with the following components: `x`, `f`, `q`, `y`, `r`, and `u`, whereas if the input object is of class `ar_conv`, this function will return a list with the following components: `acceptance_rate_cumulative`, `y`, `mean_cumulative`, and `var_cumulative`.
}
\description{
Show the data for an object of class `ar` or `ar_conv`.
}
\examples{
# pair of standard normal distribution (target) and uniform distribution (proposal)
f_norm <- "(1/sqrt(2*pi*1^2))*exp((-(x-0)^2)/(2*1^2))"
q_unif <- "runif(n, -4, 4)"
example_norm_unif <- ar(f_norm, q_unif, c=3.2, n=1000)
ar_print(example_norm_unif)

}
