% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_summary.R
\name{ar_summary}
\alias{ar_summary}
\title{ar_summary}
\usage{
ar_summary(model)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
The value of `x * 2`.
}
\description{
Multiplies a number by two using a C++ back end.
}
