% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_summary.R
\name{ar_summary}
\alias{ar_summary}
\title{Summaries of an object of class `ar`}
\usage{
ar_summary(object)
}
\arguments{
\item{object}{an object of class `ar`.}
}
\value{
The summary of an object of class `ar` with the following components: `number of iterations`, `acceptance rate`, `empirical expected value of target distribution`, `empirical variance of target distribution`, `suggested minimum c`.
}
\description{
Provide a summary for an object of class `ar`.
}
\examples{
# pair of standard normal distribution (target) and uniform distribution (proposal)
f_norm <- "(1/sqrt(2*pi*1^2))*exp((-(x-0)^2)/(2*1^2))"
q_unif <- "runif(n, -4, 4)"
example_norm_unif <- ar(f_norm, q_unif, c=3.2, n=1000)
ar_summary(example_norm_unif)

}
