% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ar}
\alias{plot.ar}
\title{Plots of an object of class `ar`.}
\usage{
\method{plot}{ar}(x, ...)
}
\arguments{
\item{x}{an object of class `ar`.}

\item{...}{additional arguments}
}
\value{
An overlay plot of the target density f(x) and proposal density q(x).
}
\description{
Display an overlay plot of the target density f(x) and proposal density q(x).
}
