% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ar_conv}
\alias{plot.ar_conv}
\title{Plots of an object of class `ar_conv`.}
\usage{
\method{plot}{ar_conv}(model, which = c(1, 2, 3, 4))
}
\arguments{
\item{model}{an object of class `ar_conv`.}

\item{...}{additional arguments}
}
\value{
Convergence plots of the sampler.
}
\description{
Display an overlay plot of the target density f(x) and proposal density q(x).
}
